+++ src/components/Buttons/EspWebButton/EspWebButton.tsx
@@ -16,11 +16,17 @@ export interface IProps {
     step?: string
     manifest: string
     checkSameFirmware: (manifest: { name: string }, improvInfo: { firmware: string }) => void
+    onClickESPButton: () => void
 }
 
 export const EspWebButton: Component<IProps> = (props) => {
     return (
-        <esp-web-install-button overrides={props.checkSameFirmware} manifest={props.manifest}>
+        <esp-web-install-button
+            overrides={props.checkSameFirmware}
+            manifest={props.manifest}
+            onClick={() => {
+                props.onClickESPButton()
+            }}>
             <button
                 slot="activate"
                 class="card card-actions !outline-none mt-0 shadow-none bg-[#192736] flex flex-col justify-between p-[14px] border-solid border-1 focus-visible:border-[#817DF7] border-[#192736] rounded-[24px] hover:border-[#817DF7] min-h-[210px] h-full max-w-[197px]  w-full ">
@@ -29,7 +35,7 @@ export const EspWebButton: Component<IProps> = (props) => {
                     <Show when={props.step}>
                         <div>
                             <p
-                                class="not-italic font-[500] text-white leading-[20px] text-[24px] text-left"
+                                class="not-italic font-normal text-white leading-[20px] text-[24px] text-left"
                                 style={{ color: 'rgba(166, 170, 211, 0.50)' }}>
                                 {props.step}
                             </p>
@@ -37,7 +43,7 @@ export const EspWebButton: Component<IProps> = (props) => {
                     </Show>
                 </div>
                 <div>
-                    <p class="not-italic font-[500] text-white leading-[20px]min-[800px] min-[800px]:text-[16px] text-[14px] text-left">
+                    <p class="not-italic font-normal text-white leading-[20px]min-[800px] min-[800px]:text-[16px] text-[14px] text-left">
                         {props.label}
                     </p>
                 </div>
@@ -50,7 +56,7 @@ export const EspWebButton: Component<IProps> = (props) => {
                     <div>{props.img}</div>
                 </div>
                 <div>
-                    <p class="not-italic font-[500] text-white leading-[20px]min-[800px] min-[800px]:text-[16px] text-[14px] text-left">
+                    <p class="not-italic font-normal text-white leading-[20px]min-[800px] min-[800px]:text-[16px] text-[14px] text-left">
                         Ah snap, your browser doesn't work!
                     </p>
                 </div>
@@ -63,7 +69,7 @@ export const EspWebButton: Component<IProps> = (props) => {
                     <div>{props.img}</div>
                 </div>
                 <div>
-                    <p class="not-italic font-[500] text-white leading-[20px]min-[800px] min-[800px]:text-[16px] text-[14px] text-left">
+                    <p class="not-italic font-normal text-white leading-[20px]min-[800px] min-[800px]:text-[16px] text-[14px] text-left">
                         Ah snap, you are not allowed to use this on HTTP!
                     </p>
                 </div>