+++ src/components/Inputs/Input/Input.tsx
@@ -17,7 +17,10 @@ const Input: Component<IProps> = (props) => {
             autocomplete={props.autoComplete}
             autofocus={props.autoFocus}
             class="h-[39px] bg-[#192736] w-full rounded-[6px] border-solid border-1 border-[#192736] placeholder-white text-[12px] text-white focus:border-[#192736] focus:ring-[#817DF7]"
-            onInput={(e) => props.onChange((e.target as HTMLInputElement).value)}
+            onInput={(e) => {
+                props.onChange(e.currentTarget.value)
+                e.currentTarget.value = props.value
+            }}
             placeholder={props.placeholder}
             type={props.type}
             name={props.id}