+++ src/containers/FlashFirmware/FlashFirmware.tsx
@@ -7,10 +7,11 @@ import { convertFileSrc } from '@tauri-apps/api/tauri'
 import { WebviewWindow, appWindow, getCurrent } from '@tauri-apps/api/window'
 import { createEffect, createMemo, createSignal, onCleanup } from 'solid-js'
 import { debug, error } from 'tauri-plugin-log-api'
+import type { INavigator, INavigatorPort } from '@interfaces/interfaces'
 import FlashFirmware from '@pages/FlashFirmware/FlashFirmware'
-import { installModalClassName, installModalTarget, installationSuccess, usb } from '@src/static'
+import { portBaudRate, usb } from '@src/static'
 import { ENotificationType, TITLEBAR_ACTION } from '@src/static/types/enums'
-import { CustomHTMLElement } from '@src/static/types/interfaces'
+import { sleep } from '@src/utils'
 import { useAppAPIContext } from '@store/context/api'
 import { useAppNotificationsContext } from '@store/context/notifications'
 
@@ -21,17 +22,34 @@ export const ManageFlashFirmware = () => {
         getFirmwareType,
         activeBoard,
         ssid,
+        mdns,
         password,
         apModeStatus,
         setAPModeStatus,
         useRequestHook,
     } = useAppAPIContext()
     const { addNotification } = useAppNotificationsContext()
-
     const [manifest, setManifest] = createSignal<string>('')
-    const [port, setPort] = createSignal<Navigator | null>(null)
-    const [installationConfirmed, setInstallationConfirmed] = createSignal<boolean>(false)
     const [response, setResponse] = createSignal<object>()
+    const [port, setPort] = createSignal<INavigatorPort | undefined>(undefined)
+    const [isSending, setIsSending] = createSignal<boolean>(false)
+
+    const config = createMemo(() => {
+        // wifi config
+        const wifiConfig = { command: 'set_wifi', data: { ssid: ssid(), password: password() } }
+        //mdns config
+        const mdnsConfig = { command: 'set_mdns', data: { hostname: mdns() } }
+
+        return JSON.stringify({ commands: [mdnsConfig, wifiConfig] })
+    })
+
+    const notify = (label: string, type: ENotificationType) => {
+        addNotification({
+            title: label,
+            message: label,
+            type: type,
+        })
+    }
 
     const erase = async () => {
         const appConfigPath = await appConfigDir()
@@ -96,6 +114,97 @@ export const ManageFlashFirmware = () => {
         await useRequestHook('save', '192.168.4.1')
     }
 
+    const validateBoardConnection = async (portSignature: INavigatorPort) => {
+        await portSignature.setSignals({
+            dataTerminalReady: false,
+            requestToSend: true,
+        })
+        await sleep(250)
+        await portSignature.setSignals({
+            dataTerminalReady: false,
+            requestToSend: false,
+        })
+        await sleep(250)
+    }
+
+    const closePort = async () => {
+        await port()?.close()
+        setPort(undefined)
+    }
+
+    const onClickUpdateNetworkSettings = async () => {
+        setIsSending(true)
+        let portSignature: INavigatorPort | undefined
+        // validate if board is still connected
+        try {
+            const serialPort = port()
+            if (serialPort) {
+                await validateBoardConnection(serialPort)
+            }
+        } catch (err) {
+            await closePort()
+        }
+
+        if (!port()) {
+            portSignature = await new Promise((resolve) => {
+                try {
+                    const port = (navigator as INavigator).serial.requestPort()
+                    resolve(port)
+                } catch {
+                    resolve(undefined)
+                }
+            })
+            setPort(portSignature)
+        } else {
+            portSignature = port()
+        }
+
+        if (!portSignature) {
+            setIsSending(false)
+            await closePort()
+            notify(
+                'Failed to open the serial port, try again or contact us on Discord.',
+                ENotificationType.INFO,
+            )
+            return
+        }
+
+        notify('preparing credentials', ENotificationType.INFO)
+        try {
+            await portSignature.open({ baudRate: portBaudRate })
+        } catch {
+            // we can ignore this error
+        }
+
+        const writableStream = (
+            portSignature as unknown as { writable: WritableStream }
+        ).writable.getWriter()
+        await writableStream.write(new TextEncoder().encode(config()))
+
+        try {
+            writableStream.releaseLock()
+        } catch {
+            // we can ignore this error
+        }
+
+        await sleep(1000)
+        notify('sending credentials', ENotificationType.INFO)
+        await sleep(4000)
+
+        setIsSending(false)
+        notify('Sent credentials', ENotificationType.INFO)
+    }
+
+    const handleCatchUpdateNetworkSettings = async (err: unknown) => {
+        setIsSending(false)
+        await closePort()
+        if (err instanceof Error) {
+            if (err.name === 'NotFoundError') return
+            notify('Failed to send wifi credentials', ENotificationType.ERROR)
+            return
+        }
+    }
+
     createEffect(() => {
         if (apModeStatus()) {
             listenToResponse().catch(console.error)
@@ -122,77 +231,9 @@ export const ManageFlashFirmware = () => {
             })
     })
 
-    const configureWifiConnection = async () => {
-        // wifi config
-        const wifiConfig = { command: 'set_wifi', data: { ssid: ssid(), password: password() } }
-
-        const writableStream = (
-            port() as unknown as { writable: WritableStream }
-        ).writable.getWriter()
-
-        const wifiConfigJSON = JSON.stringify(wifiConfig)
-        await writableStream.write(new TextEncoder().encode(wifiConfigJSON))
-        writableStream.close()
-        setInstallationConfirmed(false)
-        setPort(null)
-        addNotification({
-            title: 'WIFI configured',
-            message: 'WIFI has been configured',
-            type: ENotificationType.SUCCESS,
-        })
-    }
-
-    createEffect(() => {
-        if (isUSBBoard()) return
-        document.addEventListener('click', (e) => {
-            const targetElement = e.target as HTMLElement
-            const targetValue = targetElement.innerText
-            const className = targetElement.className
-            if (className === installModalClassName && targetValue === installModalTarget) {
-                const el: CustomHTMLElement | null = document.querySelector('[state="INSTALL"]')
-                if (el?.port) setPort(el.port)
-                return
-            }
-            if (targetValue === 'Back') {
-                setInstallationConfirmed(false)
-                setPort(null)
-            }
-        })
-    })
-
-    createEffect(() => {
-        const playInterval = port() !== null
-        const intervalId = setInterval(() => {
-            if (!playInterval) return
-            const el: HTMLElement | null = document.querySelector('[state="INSTALL"]')
-            const ewtDialog = el?.shadowRoot?.querySelector('ewt-page-message')
-            const label = ewtDialog?.getAttribute('label')
-            if (label === installationSuccess) {
-                setInstallationConfirmed(true)
-            }
-        }, 200)
-        return () => clearInterval(intervalId)
-    })
-
-    createEffect(() => {
-        const handleWifiConfigurationError = () => {
-            setInstallationConfirmed(false)
-            setPort(null)
-            addNotification({
-                title: 'WIFI configuration failed',
-                message: 'Failed to configure WIFI',
-                type: ENotificationType.ERROR,
-            })
-        }
-        if (installationConfirmed() && port() !== null) {
-            if (!apModeStatus()) {
-                configureWifiConnection().catch(handleWifiConfigurationError)
-            }
-        }
-    })
-
     return (
         <FlashFirmware
+            isSending={isSending()}
             isAPModeActive={apModeStatus()}
             isUSBBoard={isUSBBoard()}
             manifest={manifest()}
@@ -211,6 +252,17 @@ export const ManageFlashFirmware = () => {
                     default:
                         return
                 }
+            }}
+            onClickESPButton={() => {
+                closePort().catch(() => {})
+            }}
+            onClickUpdateNetworkSettings={() => {
+                onClickUpdateNetworkSettings().catch((err) => {
+                    if ((err as DOMException).name === 'NotFoundError') {
+                        alert('Failed to open the serial port, try again or contact us on Discord.')
+                        return
+                    }
+                })
             }}
             onClickConfigurAPMode={() => {
                 if (!apModeStatus()) return
@@ -244,6 +296,9 @@ export const ManageFlashFirmware = () => {
                         type: ENotificationType.ERROR,
                     })
                 })
+            }}
+            sendWifiCredentials={() => {
+                onClickUpdateNetworkSettings().catch(handleCatchUpdateNetworkSettings)
             }}
             onClickEraseSoft={() => {
                 ask('This action cannot be reverted. Are you sure?', {