+++ src/static/index.ts
@@ -1,21 +1,23 @@
-import { BOARD_TYPE, STEP_STATUS_ENUM } from './types/enums'
+import { BOARD_TYPE, CHANNEL_TYPE, STEP_STATUS_ENUM } from './types/enums'
+import { type IChannelOptions } from '@interfaces/interfaces'
 
 export const supportedBoards: string[] = [BOARD_TYPE.XIAOSENSES_3, BOARD_TYPE.XIAOSENSES_3_USB]
 export const debugModes: string[] = ['off', 'error', 'warn', 'info', 'debug', 'trace']
-export const radius = 25
+export const defaultMdnsLength = 24
+export const portBaudRate = 115200
+export const mdnsLength = 12
+export const radius = 24
 export const usb = 'USB'
-export const installModalClassName = 'mdc-button__label'
-export const installModalTarget = 'Install'
-export const installationSuccess = 'Installation complete!'
 export const questionModalId = 'questionModal'
 export const apModalID = 'apMode'
+export const wifiModalID = 'wifiMode'
 export const debugModalId = 'debugModal'
+export const staticMdns = 'openiristracker'
 
 const circleSize = Math.PI * (radius * 2)
 
 export const stepStatus: {
     [key in STEP_STATUS_ENUM]: {
-        step: string
         description: string
         dashoffset: string
         index: string
@@ -23,19 +25,16 @@ export const stepStatus: {
 } = {
     [STEP_STATUS_ENUM.SELECT_BOARD]: {
         index: '1',
-        step: 'Step 1',
         description: 'Select board',
-        dashoffset: (((100 - 0) / 100) * circleSize).toString(),
+        dashoffset: ((105 / 100) * circleSize).toString(),
     },
     [STEP_STATUS_ENUM.CONFIGURE_WIFI]: {
         index: '2',
-        step: 'Step 2',
         description: 'Configure wifi network',
-        dashoffset: (((100 - 50) / 100) * circleSize).toString(),
+        dashoffset: (((105 - 50) / 100) * circleSize).toString(),
     },
     [STEP_STATUS_ENUM.FLASH_FIRMWARE]: {
         index: '3',
-        step: 'Step 3',
         description: 'Flash firmware assets',
         dashoffset: (((100 - 100) / 100) * circleSize).toString(),
     },
@@ -55,7 +54,19 @@ export const BoardDescription: {
     [BOARD_TYPE.WROOMS_3_USB]: 'FREENOVE-ESP32-S3 (wired mode)',
     [BOARD_TYPE.WROOMS_3QIOUSB]: 'FREENOVE-ESP32-S3 (wired mode, for boards with octal flash)',
     // eslint-disable-next-line quotes
-    [BOARD_TYPE.XIAOSENSES_3]: "SeedStudio's XAIO ESP32-S3 Sense (wireless mode)",
+    [BOARD_TYPE.XIAOSENSES_3]: "SeedStudio's XIAO ESP32-S3 Sense (wireless mode)",
     // eslint-disable-next-line quotes
-    [BOARD_TYPE.XIAOSENSES_3_USB]: "SeedStudio's XAIO ESP32-S3 Sense (wired mode)",
+    [BOARD_TYPE.XIAOSENSES_3_USB]: "SeedStudio's XIAO ESP32-S3 Sense (wired mode)",
+}
+
+export const ChannelOptions: Record<CHANNEL_TYPE, IChannelOptions> = {
+    [CHANNEL_TYPE.OFFICIAL]: {
+        label: CHANNEL_TYPE.OFFICIAL,
+        description: 'Official channel for official releases.',
+    },
+    [CHANNEL_TYPE.BETA]: {
+        label: CHANNEL_TYPE.BETA,
+        description:
+            'This channel is for testing purposes only. It is not recommended for day to day usage',
+    },
 }