+++ src/static/types/interfaces.ts
@@ -1,13 +1,9 @@
 import { ENotificationAction, ENotificationType } from './enums'
-import type { RESTStatus, RESTType } from '@src/static/types/enums'
+import type { CHANNEL_TYPE, RESTStatus, RESTType } from '@src/static/types/enums'
 import type { DebugMode } from '@static/types'
 import type { WebviewWindow } from '@tauri-apps/api/window'
 import type { ToasterStore } from 'solid-headless'
 import type { JSXElement } from 'solid-js'
-
-export interface CustomHTMLElement extends HTMLElement {
-    port: Navigator
-}
 
 //* Utility Interfaces
 
@@ -85,6 +81,15 @@ export interface IGHRest {
     version: string
 }
 
+export interface IGHResponse {
+    url: string
+    status: number
+    ok: boolean
+    headers: Record<string, string>
+    rawHeaders: Record<string, string[]>
+    data: IGHRelease
+}
+
 export interface IGHRelease {
     data: object
     headers: object
@@ -92,6 +97,7 @@ export interface IGHRelease {
     ok: boolean
     status: number
     url: string
+    prerelease?: boolean
 }
 
 export interface IRestProps {
@@ -138,6 +144,8 @@ export interface AppStoreAPI {
     ssid: string
     password: string
     apModeStatus: boolean
+    mdns: string
+    channelMode: CHANNEL_TYPE
 }
 
 export interface UiStore {
@@ -145,4 +153,32 @@ export interface UiStore {
     showNotifications?: boolean
     menuOpen?: MenuOpen | null
     contextAnchor?: HTMLElement | null
+}
+
+interface SerialOutputSignals {
+    dataTerminalReady?: boolean | undefined
+    requestToSend?: boolean | undefined
+    break?: boolean | undefined
+}
+
+export interface INavigatorPort extends Navigator {
+    open: ({ baudRate }: { baudRate: number }) => Promise<void>
+    setSignals(signals: SerialOutputSignals): Promise<void>
+    close: () => Promise<void>
+}
+
+export interface INavigator extends Navigator {
+    serial: {
+        requestPort: () => INavigatorPort
+    }
+}
+
+export interface IChannelOptions {
+    label: CHANNEL_TYPE
+    description: string
+}
+
+export interface IDropdownList {
+    label: string | CHANNEL_TYPE
+    description?: string
 }