+++ src/static/types/types.ts
@@ -1,0 +1,4 @@
+export type IEventType = FocusEvent & {
+    currentTarget: HTMLDivElement
+    target: Element
+}